requirements.txt
numpy==1.26.3
opencv-python==4.8.1.78
Pillow==10.4.0
scikit-learn==1.5.2
PyMaxflow==1.3.2
matplotlib==3.7.1
torch==2.4.1
torchvision==0.19.1
timm==0.4.12
scipy==1.13.1